Extract the settings dialog and leave MainWindow a navigation shell
Move the settings dialog and the devices list out of MainWindow. A SettingsStore
mirrors macOS's ConfigStore — the current settings snapshot plus a Reload that
notifies — and the dialog renders from it, registering its config re-read on the
projection registry only while open (the last of the refresh-callback fields
MainWindow held). The devices list becomes a MembersPane. With this the
per-feature stores and dialogs own their own state and UI, and MainWindow is the
navigation shell the plan describes: window chrome, the XAML-bound handlers as
one-line delegates, library open/close/switch, and the banner/sync/status
renderers — the two dead handle-forwarders it no longer needs are gone.
